refactor(auth) : DTO 패키지 위치 변경 및 네이밍 변경
- `dto.auth.request` 및 `dto.auth.response` 패키지를 `auth.dto.request` 및 `auth.dto.response`로 이동
- `SignInRequestDto` → `SignInRequest` 로 변경하여 네이밍 통일
- `SignInResponseDto` → `SignInResponse` 로 변경 및 리팩토링
- `AuthSignInService` 인터페이스의 반환 타입을 `SignInResponseDto` → `SignInResponse` 로 수정
- `AuthService.signIn` 메서드의 반환 타입 변경
- `AccessTokenGetResponseDto` 패키지 변경 적용
- Swagger API 문서에서 변경된 DTO 반영
