+++ src/main/java/org/terning/terningserver/auth/application/reissue/AuthReissueService.java
@@ -1,6 +1,6 @@
 package org.terning.terningserver.auth.application.reissue;
 
-import org.terning.terningserver.dto.auth.response.AccessTokenGetResponseDto;
+import org.terning.terningserver.auth.dto.response.AccessTokenGetResponseDto;
 
 public interface AuthReissueService {
     AccessTokenGetResponseDto reissueToken(String refreshToken);
