+++ src/main/java/org/terning/terningserver/auth/application/signup/AuthSignUpService.java
@@ -1,8 +1,8 @@
 package org.terning.terningserver.auth.application.signup;
 
-import org.terning.terningserver.dto.auth.request.SignUpFilterRequestDto;
-import org.terning.terningserver.dto.auth.request.SignUpRequestDto;
-import org.terning.terningserver.dto.auth.response.SignUpResponseDto;
+import org.terning.terningserver.auth.dto.request.SignUpFilterRequestDto;
+import org.terning.terningserver.auth.dto.request.SignUpRequestDto;
+import org.terning.terningserver.auth.dto.response.SignUpResponseDto;
 
 public interface AuthSignUpService {
     SignUpResponseDto signUp(String authId, SignUpRequestDto request);
