+++ src/main/java/org/terning/terningserver/auth/dto/request/SignUpFilterRequestDto.java
@@ -1,7 +1,6 @@
-package org.terning.terningserver.dto.auth.request;
+package org.terning.terningserver.auth.dto.request;
 
 import lombok.Builder;
-import lombok.NonNull;
 
 import static lombok.AccessLevel.PRIVATE;
