+++ src/main/java/org/terning/terningserver/auth/dto/response/AccessTokenGetResponseDto.java
@@ -1,4 +1,4 @@
-package org.terning.terningserver.dto.auth.response;
+package org.terning.terningserver.auth.dto.response;
 
 import lombok.Builder;
 import lombok.NonNull;
