+++ src/main/java/org/terning/terningserver/controller/swagger/AuthSwagger.java
@@ -7,22 +7,22 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestHeader;
-import org.terning.terningserver.dto.auth.request.SignInRequestDto;
-import org.terning.terningserver.dto.auth.request.SignUpFilterRequestDto;
-import org.terning.terningserver.dto.auth.request.SignUpRequestDto;
-import org.terning.terningserver.dto.auth.response.AccessTokenGetResponseDto;
-import org.terning.terningserver.dto.auth.response.SignInResponseDto;
-import org.terning.terningserver.dto.auth.response.SignUpResponseDto;
+import org.terning.terningserver.auth.dto.request.SignInRequest;
+import org.terning.terningserver.auth.dto.request.SignUpFilterRequestDto;
+import org.terning.terningserver.auth.dto.request.SignUpRequestDto;
+import org.terning.terningserver.auth.dto.response.AccessTokenGetResponseDto;
+import org.terning.terningserver.auth.dto.response.SignInResponse;
+import org.terning.terningserver.auth.dto.response.SignUpResponseDto;
 import org.terning.terningserver.exception.dto.SuccessResponse;
 
 @Tag(name = "Auth", description = "소셜 로그인 및 회원가입 API")
 public interface AuthSwagger {
 
     @Operation(summary = "소셜 로그인", description = "AuthType에 맞는 소셜 로그인 API")
-    ResponseEntity<SuccessResponse<SignInResponseDto>> signIn(
+    ResponseEntity<SuccessResponse<SignInResponse>> signIn(
             @Parameter(name = "Authorization", description = "", example = "authAccessToken")
             @RequestHeader("Authorization") String authAccessToken,
-            @RequestBody SignInRequestDto request
+            @RequestBody SignInRequest request
     );
 
     @Operation(summary = "토큰 재발급", description = "토큰 재발급 API")
